# apps/backend/openapi.yaml
openapi: 3.1.0
info:
  title: Health Companion API
  version: 0.1.0
  description: API for the Health Companion application - V1 MVP
  contact:
    name: Health Companion Team
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /meals/photo:
    post:
      tags:
        - nutrition
      summary: Upload meal photo
      description: Upload a photo to create a meal entry (draft status)
      operationId: uploadMealPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
                  description: The meal photo to upload (max 8MB, jpg/png/heic)
                takenAt:
                  type: string
                  format: date-time
                  description: When the photo was taken (ISO datetime)
                  example: '2025-08-30T12:30:00Z'
                notes:
                  type: string
                  maxLength: 500
                  description: Optional notes about the meal
                  example: 'Lunch at the office'
              required:
                - image
      responses:
        '201':
          description: Meal created successfully with photo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealPhotoResponse'
        '400':
          description: Invalid file, missing image, or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload too large (file size exceeds 8MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /meals/{mealId}:
    get:
      tags:
        - nutrition
      summary: Get meal by ID
      description: Retrieve a specific meal by its ID
      operationId: getMealById
      parameters:
        - name: mealId
          in: path
          required: true
          description: The ID of the meal to retrieve
          schema:
            type: string
            format: uuid
            example: '123e4567-e89b-12d3-a456-426614174000'
      responses:
        '200':
          description: Meal retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meal'
        '404':
          description: Meal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /hydration:
    get:
      tags:
        - hydration
      summary: Get hydration entries
      description: Retrieve hydration entries for a specific date
      parameters:
        - name: date
          in: query
          description: Date to filter hydration entries (YYYY-MM-DD format)
          schema:
            type: string
            format: date
            example: '2025-08-30'
      responses:
        '200':
          description: Hydration entries retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HydrationResponse'
    post:
      tags:
        - hydration
      summary: Log hydration
      description: Log a new hydration entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHydrationRequest'
      responses:
        '201':
          description: Hydration logged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HydrationEntry'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input provided
          required:
            - code
            - message
      required:
        - error
...
    Meal:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        name:
          type: string
          example: Grilled Chicken Salad
        type:
          type: string
          enum: [breakfast, lunch, dinner, snack]
          example: lunch
        takenAt:
          type: string
          format: date-time
          example: '2025-08-30T12:30:00Z'
        status:
          type: string
          enum: [draft, recognized, final]
          example: final
        totalCalories:
          type: number
          minimum: 0
          example: 450
        totalProteinGrams:
          type: number
          minimum: 0
          example: 35
        totalCarbsGrams:
          type: number
          minimum: 0
          example: 15
        totalFatGrams:
          type: number
          minimum: 0
          example: 20
        photoUrl:
          type: string
          format: uri
          example: https://example.com/photos/meal-123.jpg
        items:
          type: array
          items:
            $ref: '#/components/schemas/FoodItem'
        createdAt:
          type: string
          format: date-time
          example: '2025-08-30T12:30:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2025-08-30T12:30:00Z'
      required:
        - id
        - name
        - type
        - takenAt
        - status
        - totalCalories
        - createdAt
        - updatedAt
    FoodItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        name:
          type: string
          example: Grilled Chicken Breast
        quantity:
          type: number
          minimum: 0
          example: 150
        unit:
          type: string
          example: grams
        calories:
          type: number
          minimum: 0
          example: 165
        proteinGrams:
          type: number
          minimum: 0
          example: 31
        carbsGrams:
          type: number
          minimum: 0
          example: 0
        fatGrams:
          type: number
          minimum: 0
          example: 3.6
        confidence:
          type: number
          minimum: 0
          maximum: 1
          example: 0.95
          description: AI confidence score for recognized items
      required:
        - id
        - name
        - quantity
        - unit
        - calories
...
    MealPhotoResponse:
      type: object
      properties:
        mealId:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
          description: The ID of the created meal
        photoUrl:
          type: string
          format: uri
          example: 'https://cdn.example.com/meals/meal-uuid.jpg'
          description: URL of the uploaded photo
        status:
          type: string
          enum: [draft]
          example: draft
          description: Status of the meal (always draft for new uploads)
        createdAt:
          type: string
          format: date-time
          example: '2025-08-30T12:30:00Z'
          description: When the meal was created
      required:
        - mealId
        - photoUrl
        - status
        - createdAt
...
    HydrationEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '123e4567-e89b-12d3-a456-426614174000'
        amountMl:
          type: number
          minimum: 0
          maximum: 3000
          example: 250
        type:
          type: string
          enum: [water, tea, coffee, juice, other]
          example: water
        takenAt:
          type: string
          format: date-time
          example: '2025-08-30T14:30:00Z'
        createdAt:
          type: string
          format: date-time
          example: '2025-08-30T14:30:00Z'
      required:
        - id
        - amountMl
        - type
        - takenAt
        - createdAt
    CreateHydrationRequest:
      type: object
      properties:
        amountMl:
          type: number
          minimum: 0
          maximum: 3000
          example: 250
        type:
          type: string
          enum: [water, tea, coffee, juice, other]
          example: water
        takenAt:
          type: string
          format: date-time
          example: '2025-08-30T14:30:00Z'
      required:
        - amountMl
        - type
        - takenAt
    HydrationResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/HydrationEntry'
        totalMl:
          type: number
          minimum: 0
          example: 1500
        goalMl:
          type: number
          minimum: 0
          example: 2500
        date:
          type: string
          format: date
          example: '2025-08-30'
      required:
        - entries
        - totalMl
        - goalMl
        - date
...